This is the basic version of the  PMPlayer.
This file includes the code for the music player. It is a python file. The requirements to run this file are none. but  Python3 is required to be installed on the machine. It uses many of the Python modules such as Tkinter and pygame and mutagen. These modules helped me build this music player.